+++ config/autoload/preview-generation.global.php
@@ -1,0 +1,8 @@
+<?php
+return [
+
+    'preview_generation' => [
+        'files_location' => 'data/cache',
+    ],
+
+];