+++ module/CLI/config/cli.config.php
@@ -4,11 +4,13 @@ use Shlinkio\Shlink\CLI\Command;
 return [
 
     'cli' => [
+        'locale' => env('CLI_LOCALE', 'en'),
         'commands' => [
             Command\Shortcode\GenerateShortcodeCommand::class,
             Command\Shortcode\ResolveUrlCommand::class,
             Command\Shortcode\ListShortcodesCommand::class,
             Command\Shortcode\GetVisitsCommand::class,
+            Command\Shortcode\GeneratePreviewCommand::class,
             Command\Visit\ProcessVisitsCommand::class,
             Command\Config\GenerateCharsetCommand::class,
             Command\Config\GenerateSecretCommand::class,