+++ module/Common/src/ErrorHandler/ErrorHandlerManagerInterface.php
@@ -1,9 +1,0 @@
-<?php
-namespace Shlinkio\Shlink\Common\ErrorHandler;
-
-use Interop\Container\ContainerInterface;
-
-interface ErrorHandlerManagerInterface extends ContainerInterface
-{
-
-}