+++ module/Common/src/Image/ImageBuilderInterface.php
@@ -1,0 +1,8 @@
+<?php
+namespace Shlinkio\Shlink\Common\Image;
+
+use Zend\ServiceManager\ServiceLocatorInterface;
+
+interface ImageBuilderInterface extends ServiceLocatorInterface
+{
+}