+++ module/Core/src/Service/ShortUrlServiceInterface.php
@@ -2,6 +2,7 @@
 namespace Shlinkio\Shlink\Core\Service;
 
 use Shlinkio\Shlink\Core\Entity\ShortUrl;
+use Shlinkio\Shlink\Core\Exception\InvalidShortCodeException;
 use Zend\Paginator\Paginator;
 
 interface ShortUrlServiceInterface
@@ -11,4 +12,12 @@ interface ShortUrlServiceInterface
      * @return ShortUrl[]|Paginator
      */
     public function listShortUrls($page = 1);
+
+    /**
+     * @param string $shortCode
+     * @param string[] $tags
+     * @return ShortUrl
+     * @throws InvalidShortCodeException
+     */
+    public function setTagsByShortCode($shortCode, array $tags = []);
 }