+++ module/Rest/config/rest.config.php
@@ -1,9 +1,0 @@
-<?php
-return [
-
-    'rest' => [
-        'username' => env('REST_USER'),
-        'password' => env('REST_PASSWORD'),
-    ],
-
-];